move typecheck, remove unused export
